// src/models/Package.d.ts
export type PackageManagerType = 'npm' | 'yarn' | 'pnpm';